requirements.txt
pyyaml==6.0.1
selenium==4.9.1
robotframework==6.01